Move ConfigWatchers into galaxy.webapps.galaxy.config_watchers
